gh-13: moved command into cmd folder and added use on cobra for command management
