+++ cmd/fail.go
@@ -1,0 +1,11 @@
+package cmd
+
+import (
+	"github.com/sirupsen/logrus"
+	"os"
+)
+
+func fail(message string, data ...interface{}) {
+	logrus.Errorf("[main] "+message, data...)
+	os.Exit(1)
+}
